+++ tests/ui/no_comma_after_eq.rs
@@ -1,0 +1,10 @@
+use dst_factory::make_dst_factory;
+
+#[make_dst_factory(create, debug, eq no_std)]
+struct NoCommaAfterEq {
+    id: i32,
+    data: str,
+}
+
+fn main() {
+}