+++ tests/ui/no_comma_after_ord.rs
@@ -1,0 +1,10 @@
+use dst_factory::make_dst_factory;
+
+#[make_dst_factory(create, eq, ord no_std)]
+struct NoCommaAfterOrd {
+    id: i32,
+    data: str,
+}
+
+fn main() {
+}